Moved backend test folder to match structure of the frontend
